+++ config/myjob.php
@@ -9,16 +9,16 @@ return [
 		'numberDisplaySearch' => 24,
 	],
 	'routes'    => [ // should contains the same real pages as routes.php (i.e. no-redirect)
-		'home'       => 'PublicController@index',
+		'home'       => 'HomeController@index',
 		'jobs'       => 'AdController@index',
 		'myjobs'     => 'AdController@created',
 		'newjob'     => 'AdController@create',
 		'editjob'    => 'AdController@edit',
 		'moderation' => 'ModerationController@adsToModerate',
 		'options'    => 'OptionsController@index',
-		'help'       => 'PublicController@help',
-		'connect'    => 'PublicController@connect',
-		'disconnect' => 'PublicController@disconnect',
+		'help'       => 'HelpController@index',
+		'connect'    => 'HomeController@connect',
+		'disconnect' => 'HomeController@disconnect',
 	],
 
 ];
