+++ resources/lang/fr/general.php
@@ -100,10 +100,8 @@ return [
 		'description'   => "L'<strong>Association Générale des Etudiants</strong> de l'<strong>Ecole Polytechnique Fédérale de Lausanne</strong> met gratuitement cette plateforme à disposition pour mettre en relation étudiants et employeurs. Le but est de permettre aux étudiants de trouver facilement un job/emploi <strong>durant leurs études</strong>, et en même temps d'offrir aux employeurs des profils variés et qualifiés.",
 		'noinscription' => "Aucune inscription, gestion par <strong>email</strong>.",
 		'tequila'       => "Accès <strong><em>Tequila</em></strong> requis",
-		'peopleepfl'    => "Raccourci collaborateur EPFL.",
 		'oldad'         => "Retrouver une ancienne annonce.",
 		'newpassword'   => "Recevoir un nouveau mot de passe.",
-		'outsideepfl'   => "Informations étudiants hors-EPFL.",
 		'advices'       => "Par le service académique, lorem ipsum dolor sit amet, consectetur adipiscing elit. Ut maximus, lacus ac volutpat dictum, arcu urna venenatis tellus, in tempus arcu ligula a ante. Mauris mattis, lorem sed pharetra lacinia, justo tellus ultricies purus, ac placerat augue ante et quam. Quisque id pretium turpis.",
 		'news'          => "Bienvenue sur la nouvelle version ! Aide-nous à améliorer cette grande mise à jour en signalant tout ce qui te parait anormal ou peu pratique. On attend avec impatience tes remarques et suggestions !",
 		'forgotten-link'=> "Nous vous enverrons un email contenant un nouveau lien secret de connexion. Votre lien précédent sera invalidé.",
