+++ resources/views/ads/elements/news.blade.php
@@ -1,6 +1,6 @@
 <h3 class="ui header">{{ trans('general.titles.news') }}</h3>
 <p>{{ trans('general.texts.news') }}</p>
-<a class="ui small red icon button mt" href="{{ action('PublicController@help') }}">
+<a class="ui small red icon button mt" href="{{ action('HelpController@index') }}">
     <i class="write icon"></i>
     {{ trans('general.buttons.feedback') }}
 </a>
