+++ resources/views/options/elements/notifications.blade.php
@@ -1,6 +1,6 @@
 <h3 class="ui header">{{ trans('general.titles.notifications') }}</h3>
     <p>{{ trans('general.texts.notifications') }}</p>
-    <a class="ui small red button mt" href="{{ action('PublicController@help') }}">
+    <a class="ui small red button mt" href="{{ action('HelpController@index') }}">
         <i class="help icon"></i>
         {{ trans('general.buttons.ask') }}
     </a>
