+++ resources/views/publishers/link.blade.php
@@ -1,4 +1,4 @@
-@extends('layout.layout')
+@extends('layout')
 
 @section('content')
     <div class="row">
@@ -6,10 +6,9 @@
             <h2>{{ trans('general.titles.forgotten-link') }}</h2>
             <p>{{ trans('general.texts.forgotten-link') }}</p>
             {!! Form::open([
-                'action' => 'PublicController@postForgottenLink',
+                'action' => 'PublishersController@postForgottenLink',
                 'class' => 'ui form',
             ]) !!}
-
             @if ($errors->any())
                 <div class="ui error visible message">
                     <ul class="list">
@@ -19,7 +18,6 @@
                     </ul>
                 </div>
             @endif
-
             <div class="field">
                 <input type="text" name="email" placeholder="{{ trans('general.placeholders.forgotten-link-mail') }}">
             </div>
@@ -27,7 +25,6 @@
                 <button class="ui red submit button" type="submit">{{ trans('general.buttons.submit.send-short') }}</button>
             </div>
             {!! Form::close() !!}
-
         </div>
     </div>
 @stop
